requirements_env.txt
astroid==2.0.4
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
html5lib==1.0.1
idna==2.5
isodate==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pagarme-python==3.3.0
pdfkit==0.6.0
Pillow==6.0.0
pycparser==2.19
pylint==2.1.1
Pyphen==0.9.5
pytz==2018.5
requests==2.18.2
six==1.11.0
tinycss2==1.0.2
typed-ast==1.1.0
urllib3==1.22
webencodings==0.5.1
wrapt==1.10.11
